Replace events with notifications throughout framework and documentation
