+++ OpenCqrs.sln
@@ -37,7 +37,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "docs", "docs", "{5442F24B-4
 		docs\Basics.md = docs\Basics.md
 		docs\Commands.md = docs\Commands.md
 		docs\Configuration.md = docs\Configuration.md
-		docs\Events.md = docs\Events.md
 		docs\index.md = docs\index.md
 		docs\Installation.md = docs\Installation.md
 		docs\Queries.md = docs\Queries.md
@@ -46,6 +45,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "docs", "docs", "{5442F24B-4
 		docs\Entity-Framework-Core.md = docs\Entity-Framework-Core.md
 		docs\Cosmos.md = docs\Cosmos.md
 		docs\Domain.md = docs\Domain.md
+		docs\Notifications.md = docs\Notifications.md
 	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenCqrs.Examples.EventSourcing.EntityFrameworkCore", "examples\OpenCqrs.Examples.EventSourcing.EntityFrameworkCore\OpenCqrs.Examples.EventSourcing.EntityFrameworkCore.csproj", "{6379279E-932F-41F3-BEFD-7FC432FB7E1F}"
