+++ docs/Basics.md
@@ -1,10 +1,10 @@
 # Basics
 
-The **IDispatcher** interface contains all the methods needed to send commands, publish events and get results from queries.
+The **IDispatcher** interface contains all the methods needed to send commands, get results from queries, and publish notifications.
 
 There are three kinds of requests that can be sent through the dispatcher:
 - [Commands](Commands) (single handler)
 - [Queries](Queries) (single handler)
-- [Events](Events) (multiple handlers)
+- [Notifications](Notifications) (multiple handlers)
 
-OpenCQRS uses the result pattern to return the result of commands and queries. The result contains information about the success or failure of the operation, any errors that occurred, and the data returned by the operation.
+OpenCQRS uses the result pattern to return the result of commands, queries, and notifications. The result contains information about the success or failure of the operation, any errors that occurred, and the data returned by the operation.
