+++ docs/Release-Notes.md
@@ -1,6 +1,10 @@
 # Release Notes
 
+## OpenCQRS 7.0.0-beta.2
+_**Released 26/08/2025**_
+- Replace events with notifications
+
 ## OpenCQRS 7.0.0-beta.1 
-_**Released 31/08/2025**_
+_**Released 25/08/2025**_
 - Complete rewrite of the framework
 - Upgrade to .NET 9
