+++ src/OpenCqrs/Notifications/INotification.cs
@@ -1,0 +1,3 @@
+﻿namespace OpenCqrs.Notifications;
+
+public interface INotification;
