+++ src/OpenCqrs/Notifications/IPublisher.cs
@@ -1,0 +1,8 @@
+﻿using OpenCqrs.Results;
+
+namespace OpenCqrs.Notifications;
+
+public interface IPublisher
+{
+    Task<IEnumerable<Result>> Publish<TNotification>(TNotification notification, CancellationToken cancellationToken = default) where TNotification : INotification;
+}
